# api/openapi.yaml
openapi: 3.0.3
info:
  title: Host API
  version: 0.1.0
  description: API for experiment hosting platform with jsPsych, DataPipe, and Prolific integration
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.yoursite.com
    description: Production server
paths:
...
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
              required: [email, password]
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  email: { type: string }
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: 'JWT token' }
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /experiments:
    get:
      tags: [Experiments]
      summary: List all experiments for authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of experiments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Experiment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags: [Experiments]
      summary: Create a new experiment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                description:
                  type: string
                  nullable: true
              required: [title]
      responses:
        '201':
          description: Experiment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /dashboard/overview:
    get:
      tags: [Dashboard]
      summary: Get dashboard overview statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard overview data
          content:
            application/json:
              schema:
                type: object
                properties:
                  statistics:
                    type: object
                    properties:
                      experiments:
                        type: object
                        properties:
                          total: { type: number }
                          live: { type: number }
                          with_data: { type: number }
                      participants:
                        type: object
                        properties:
                          total: { type: number }
                          this_week: { type: number }
                      data_sync:
                        type: object
                        properties:
                          synced: { type: number }
                          pending: { type: number }
                  recent_activity:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        title: { type: string }
                        live: { type: boolean }
                        created_at: { type: string, format: date-time }
                        participant_count: { type: number }
  /dashboard/experiments:
    get:
      tags: [Dashboard]
      summary: Get detailed experiment statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Detailed experiment statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: number }
                  experiments:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        title: { type: string }
                        live: { type: boolean }
                        created_at: { type: string, format: date-time }
                        statistics:
                          type: object
                          properties:
                            total_participants: { type: number }
                            synced_participants: { type: number }
                            pending_sync: { type: number }
                            last_participant: { type: string, format: date-time, nullable: true }
                        configuration:
                          type: object
                          properties:
                            has_files: { type: boolean }
                            has_prolific: { type: boolean }
                            has_datapipe: { type: boolean }
  /dashboard/activity:
    get:
      tags: [Dashboard]
      summary: Get activity timeline
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 30
          description: Number of days to look back
      responses:
        '200':
          description: Activity timeline data
          content:
            application/json:
              schema:
                type: object
                properties:
                  time_range:
                    type: object
                    properties:
                      start: { type: string, format: date-time }
                      end: { type: string, format: date-time }
                      days: { type: number }
                  daily_participants:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, format: date }
                        count: { type: number }
                  daily_experiments:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, format: date }
                        count: { type: number }
  /dashboard/storage:
    get:
      tags: [Dashboard]
      summary: Get storage usage statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Storage usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_storage:
                    type: object
                    properties:
                      bytes: { type: number }
                      mb: { type: string }
                      gb: { type: string }
                  total_files: { type: number }
                  experiments:
                    type: array
                    items:
                      type: object
                      properties:
                        experiment_id: { type: string, format: uuid }
                        experiment_title: { type: string }
                        size_bytes: { type: number }
                        size_mb: { type: string }
                        file_count: { type: number }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    Experiment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        live:
          type: boolean
          default: false
        experiment_files_path:
          type: string
          nullable: true
        prolific_study_id:
          type: string
          nullable: true
        prolific_status:
          type: string
          enum: [draft, published, completed]
          default: draft
        completion_code:
          type: string
          nullable: true
        datapipe_project_id:
          type: string
          nullable: true
        datapipe_component_id:
          type: string
          nullable: true
        datapipe_experiment_id:
          type: string
          nullable: true
        auto_delete_days:
          type: integer
          default: 30
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              msg:
                type: string
              path:
                type: string
              location:
                type: string
...